Problem: bottom tab covers some content of the tabs.
Solution: add a padding bottom to the content in layout
