+++ src/components/Discover/index.js
@@ -6,7 +6,6 @@ import AppBar from '@material-ui/core/AppBar'
 import Tabs from '@material-ui/core/Tabs'
 import Tab from '@material-ui/core/Tab'
 import Typography from '@material-ui/core/Typography'
-import Box from '@material-ui/core/Box'
 import ChatList from 'components/ChatList'
 
 function TabPanel(props) {
@@ -17,11 +16,12 @@ function TabPanel(props) {
       component="div"
       role="tabpanel"
       hidden={value !== index}
-      id={`full-width-tabpanel-${index}`}
-      aria-labelledby={`full-width-tab-${index}`}
+      id={`scrollable-auto-tabpanel-${index}`}
+      aria-labelledby={`scrollable-auto-tab-${index}`}
+      noWrap
       {...other}
     >
-      <Box p={3}>{children}</Box>
+      {children}
     </Typography>
   )
 }
