+++ src/components/Layout/index.jsx
@@ -5,7 +5,7 @@ import GlobalStyles from 'util/style/GlobalStyles'
 
 import Header from 'components/Header'
 import Footer from 'components/Footer'
-
+import Paper from '@material-ui/core/Paper'
 import config from '../../../config/meta'
 
 const Layout = ({ children }) => (
@@ -13,7 +13,9 @@ const Layout = ({ children }) => (
     <>
       <GlobalStyles />
       <Header siteTitle={config.siteTitle || config.siteTitleAlt} />
-      {children}
+      <Paper square style={{ paddingBottom: 50 }}>
+        {children}
+      </Paper>
       <Footer />
     </>
   </ThemeProvider>
