+++ src/pages/chat.jsx
@@ -1,12 +1,11 @@
 import React from 'react'
 import Layout from 'components/Layout'
 import SEO from 'components/SEO'
-import Chat from 'components/Chat'
 
 const SecondPage = () => (
   <Layout>
     <SEO title="Chat" />
-    <Chat />
+    {/* <Chat /> */}
   </Layout>
 )
